+++ packages/akashic-cli-serve/src/server/route/ApiRoute.ts
@@ -14,10 +14,10 @@ import {
 	createHandlerToDeleteRunner,
 	createHandlerToPatchRunner
 } from "../controller/RunnerController";
-import { createHandlerToGetSandboxConfig } from "../controller/SandboxConfigController";
 import { PlayStore } from "../domain/PlayStore";
 import { RunnerStore } from "../domain/RunnerStore";
 import { SocketIOAMFlowManager } from "../domain/SocketIOAMFlowManager";
+import { createHandlerToGetSandboxConfig } from "../controller/SandboxConfigController";
 
 export interface ApiRouterParameterObject {
 	targetDir: string;