+++ packages/akashic-cli-serve/src/server/route/ConfigRoute.ts
@@ -1,5 +1,5 @@
 import * as express from "express";
-import { createHandlerToGetEngineConfig } from "../controller/ConfigController";
+import { createHandlerToGetEngineConfig, handleToGetStartupOptions } from "../controller/ConfigController";
 
 export interface ConfigRouterParameterObject {
 	targetDir: string;
@@ -12,6 +12,7 @@ export const createConfigRouter = (params: ConfigRouterParameterObject): express
 	// /engineとの相違点はスクリプトアセット加工前のコンテンツを含む情報を投げること
 	// サーバー側でインスタンスを立ち上げる時は加工前のスクリプトアセットを参照する必要がある
 	configRouter.get("/content.raw.json", createHandlerToGetEngineConfig(params.targetDir, true));
+	configRouter.get("/options", handleToGetStartupOptions);
 
 	return configRouter;
 };