+++ packages/akashic-cli-export-html/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.9](https://github-com-akashic-cli/akashic-games/akashic-cli/compare/@akashic/akashic-cli-export-html@0.2.8...@akashic/akashic-cli-export-html@0.2.9) (2019-03-06)
+
+**Note:** Version bump only for package @akashic/akashic-cli-export-html
+
+
+
+
+
 ## [0.2.8](https://github-com-akashic-cli/akashic-games/akashic-cli/compare/@akashic/akashic-cli-export-html@0.2.7...@akashic/akashic-cli-export-html@0.2.8) (2019-03-04)