+++ packages/akashic-cli-modify/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.4](https://github-com-akashic-cli/akashic-games/akashic-cli/compare/@akashic/akashic-cli-modify@0.1.3...@akashic/akashic-cli-modify@0.1.4) (2019-03-06)
+
+**Note:** Version bump only for package @akashic/akashic-cli-modify
+
+
+
+
+
 ## [0.1.3](https://github-com-akashic-cli/akashic-games/akashic-cli/compare/@akashic/akashic-cli-modify@0.1.2...@akashic/akashic-cli-modify@0.1.3) (2019-03-04)